+++ Backend/src/routes/account.ts
@@ -1,4 +1,4 @@
-import {comparePWD, hashPWD, createTOTP, getSecret, getBackupCodes} from '../crypto';
+import {comparePWD, hashPWD, createTOTP, getSecret, getRecovery} from '../crypto';
 import {BodyField, QueryResult, User} from '../types';
 import {msg, attemptQuery} from '../utils';
 import {sql} from '../connection';
@@ -16,8 +16,8 @@ export default new Elysia({prefix: '/account'})
   })
   .onBeforeHandle(({user, path}) => {
     const relativePath = path.slice(8);
-    const notLog = ['/login', '/login-otp', '/login-backup', '/signup', '/signup-otp', '/signup-backup', '/signup-create'];
-    const mustLog = ['/settings', '/otp', '/full'];
+    const notLog = ['/login', '/login-otp', '/login-recovery', '/signup', '/signup-otp', '/signup-recovery', '/signup-create'];
+    const mustLog = ['/delete-otp', '/delete-full'];
 
     if (notLog.some((p) => relativePath === p) && user) {
       return msg('You are already logged in', 'info');
@@ -39,33 +39,32 @@ export default new Elysia({prefix: '/account'})
     const isPasswordCorrect = await comparePWD(password, hashedPassword);
     if (!isPasswordCorrect) return msg('Invalid credentials. Please try again', 'alert');
 
-    const has2fa = result[0].otp_auth;
+    const has2fa = result[0].totp;
     if (has2fa) return {username};
 
     const cookieValue = await jwt.sign({password, username});
     return {cookie: cookieValue};
   })
   .post('/login-otp', async ({body, jwt}) => {
-    const {code, username, password, err} = check(body as BodyField, ['code', 'username', 'password']);
+    const {token, username, password, err} = check(body as BodyField, ['token', 'username', 'password']);
     if (err) return msg(err, 'alert');
 
-    const secret = (await attemptQuery(sql`SELECT otp_auth FROM users WHERE username = ${username}`)) as QueryResult[];
-    const totp = createTOTP(secret[0].otp_auth, username);
-    const isValid = totp.generate() === code;
+    const secret = (await attemptQuery(sql`SELECT totp FROM users WHERE username = ${username}`)) as QueryResult[];
+    const totp = createTOTP(secret[0].totp, username);
 
-    if (!isValid) return msg('Incorrect validation code. Please try again', 'alert');
+    const isValid = totp.generate() === token;
+    if (!isValid) return msg('Incorrect validation token. Please try again', 'alert');
 
     const cookieValue = await jwt.sign({username, password});
     return {cookie: cookieValue};
   })
-  .post('/login-backup', async ({body, jwt}) => {
-    const {backup, username, password, err} = check(body as BodyField, ['backup', 'username', 'password']);
+  .post('/login-recovery', async ({body, jwt}) => {
+    const {code, username, password, err} = check(body as BodyField, ['code', 'username', 'password']);
     if (err) return msg(err, 'alert');
 
-    const allBackups = (await attemptQuery(sql`SELECT backup_codes FROM users WHERE username = ${username}`)) as QueryResult[];
-    const isValid = allBackups[0].backup_codes.some((b) => b === backup);
-
-    if (!isValid) return msg('Incorrect backup code. Try another one', 'alert');
+    const result = (await attemptQuery(sql`SELECT recovery FROM users WHERE username = ${username}`)) as QueryResult[];
+    const isValid = result[0].recovery.some((b) => b === code);
+    if (!isValid) return msg('Incorrect recovery code. Try another one', 'alert');
 
     const cookieValue = await jwt.sign({username, password});
     return {cookie: cookieValue};
@@ -79,27 +78,26 @@ export default new Elysia({prefix: '/account'})
 
     return {username};
   })
-  .post('/signup-otp', async ({user}) => {
+  .post('/signup-otp', async () => {
     const secret = getSecret();
-    const totp = createTOTP(secret, user!.username);
+    const totp = createTOTP(secret, 'temporarily');
     const uri = totp.toString();
     return {uri, secret};
   })
-  .post('/signup-backup', async ({body, user}) => {
-    const {code, secret, err} = check(body as BodyField, ['secret', 'code']);
+  .post('/signup-recovery', async ({body}) => {
+    const {token, secret, err} = check(body as BodyField, ['secret', 'token']);
     if (err) return msg(err, 'alert');
 
-    const totp = createTOTP(secret, user!.username);
-    const isValid = totp.generate() === code;
+    const totp = createTOTP(secret, 'temporarily');
+    const isValid = totp.generate() === token;
+    if (!isValid) return msg('Incorrect validation token. Please try again', 'alert');
 
-    if (!isValid) return msg('Incorrect validation code. Please try again', 'alert');
-
-    const backupCodes = getBackupCodes();
-    return {backup: backupCodes};
+    const recovery = getRecovery();
+    return {recovery};
   })
   .post('/signup-create', async ({jwt, body}) => {
-    const fields = ['username', 'password', '?secret', '?backups'];
-    const {password, username, secret, backups, err} = check(body as BodyField, fields);
+    const fields = ['username', 'password', '?secret', '?recovery'];
+    const {password, username, secret, recovery, err} = check(body as BodyField, fields);
     if (err) return msg(err, 'alert');
 
     const isTaken = (await attemptQuery(sql`SELECT * FROM users WHERE username = ${username}`)) as QueryResult[];
@@ -108,25 +106,23 @@ export default new Elysia({prefix: '/account'})
     const newPassword = await hashPWD(password);
     await attemptQuery(sql`INSERT INTO users (username, password) VALUES (${username}, ${newPassword})`);
 
-    if (secret && backups) {
-      await attemptQuery(sql`UPDATE users SET otp_auth = ${secret} WHERE username = ${username}`);
-      await attemptQuery(sql`UPDATE users SET backup_codes = ${backups} WHERE username = ${username}`);
+    if (secret && recovery) {
+      await attemptQuery(sql`UPDATE users SET totp = ${secret} WHERE username = ${username}`);
+      await attemptQuery(sql`UPDATE users SET recovery = ${recovery} WHERE username = ${username}`);
     }
 
     const cookieValue = await jwt.sign({password, username});
     return {cookie: cookieValue};
   })
-  .delete('/full', async ({user}) => {
-    const costumer = attemptQuery(sql`DELETE FROM users WHERE username = ${user!.username}`);
-    if (!user) return msg('Something went wrong', 'alert');
+  .delete('/delete-full', async ({user}) => {
+    const client = attemptQuery(sql`DELETE FROM users WHERE username = ${user!.username}`);
+    if (!client) return msg('Something went wrong', 'alert');
 
-    return costumer;
+    return client;
   })
-  .delete('/otp', async ({user}) => {
-    const customer = (await attemptQuery(
-      sql`UPDATE users SET otp_auth = NULL, backup_codes = NULL WHERE username = ${user!.username}`,
-    )) as QueryResult[];
-    if (!customer) return msg('Something went wrong', 'alert');
+  .delete('/delete-otp', async ({user}) => {
+    const client = await attemptQuery(sql`UPDATE users SET totp = NULL, recovery = NULL WHERE username = ${user!.username}`);
+    if (!client) return msg('Something went wrong', 'alert');
 
-    return customer;
+    return client;
   });